+++ pyechelle/hdfbuilder.py
@@ -180,38 +180,3 @@ class HDFBuilder(contextlib.ExitStack):
         fiber_group.create_dataset(
             name=f"order{abs(order)}", data=array, shape=np.shape(array)
         )
-
-
-if __name__ == "__main__":
-    from pyechelle.spectrograph import InteractiveZEMAX
-
-    # Open Link to a standalone OpticStudio instance
-    zmx = InteractiveZEMAX(name='AWESOME_SPECTROGRAPH', zemax_filepath="PATH_TO_ZMX/ZOS_FILE")
-
-    # set basic grating specifications
-    zmx.set_grating(surface='echelle', blaze=-75.964, theta=0., gamma=-0.75)
-
-    # add CCD information (only one CCD supported so far. So for instruments with multiple CCDs, you have to generate
-    # separate models for now.
-    zmx.add_ccd(1, CCD(10560, 10560, pixelsize=9))
-
-    # Add here so many fiber/fields as you wish. You don't have to fiddle with the fields in OpticStudio. The
-    # existing fields will be ignored/deleted.
-    zmx.add_field(0., 0., 75 * 4, 75, shape='rectangular', name='Science fiber')
-
-    # Add here a list with the diffraction orders you want to include
-    zmx.set_orders(1, 1, list(range(65, 100)))
-
-    # Adjust settings for the Huygens PSF. Best to check out 'reasonable' parameters manually in ZEMAX first.
-    zmx.psf_settings(image_delta=0.4, image_sampling="64x64", pupil_sampling="128x128")
-
-    # at this point, you can interact with the spectrograph interactively, e.g. by doing something like:
-    # zmx.get_psf(wavelength=0.72, order=85, fiber=1, ccd_index=1)
-    # zmx.get_wavelength_range(order=85)
-    # zmx.get_transformation(wavelength=0.72, order=85, fiber=1, ccd_index=1)
-    # and it will/should pull the appropriate values from ZEMAX
-
-    # To generate an .HDF model file, you do:
-    hdf = HDFBuilder(zmx, '../pyechelle/models/AWESOME_SPECTROGRAPH.hdf')
-    # this will take a long time...
-    hdf.save_to_hdf(n_transformation_per_order=50, n_psfs_per_order=5)