+++ tests/test_slits.py
@@ -1,26 +1,18 @@
 import inspect
 import random
 
-from numba.core.registry import CPUDispatcher
-from numba.cuda.dispatcher import CUDADispatcher
-
 import pyechelle.slit
 
-available_slits = [m[0] for m in inspect.getmembers(pyechelle.slit) if isinstance(m[1], CPUDispatcher)]
-available_cuda_slits = [m[0] for m in inspect.getmembers(pyechelle.slit) if isinstance(m[1], CUDADispatcher)]
+available_slits = [f for n, f in vars(pyechelle.slit).items() if inspect.isfunction(f) if n != 'njit']
 
 
-def test_slits_numba():
-    for s in available_slits:
-        slit = getattr(pyechelle.slit, s)
-        x, y = slit(random.random(), random.random())
+# available_cuda_slits = [m[0] for m in inspect.getmembers(pyechelle.slit) if isinstance(m[1], CUDADispatcher)]
+
+
+def test_slits():
+    for slit_func in available_slits:
+        x, y = slit_func(random.random(), random.random())
         assert 0. <= x <= 1.0
         assert 0. <= y <= 1.0
 
-
-def test_slits_native():
-    for s in available_slits:
-        slit = getattr(pyechelle.slit, s)
-        x, y = slit.py_func(random.random(), random.random())
-        assert 0. <= x <= 1.0
-        assert 0. <= y <= 1.0
+# TODO: add test for cuda slits